phase3-chatbot/requirements-pydantic-v1.txt
# Phase 3 Dependencies with Pydantic V1 (No Rust)

# Pydantic V1 (pure Python, no Rust)
pydantic==1.10.13

# FastAPI (compatible with Pydantic V1)
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# OpenAI SDK
openai>=1.10.0

# HTTP client
httpx>=0.26.0

# Environment variables
python-dotenv==1.0.0

# Authentication (JWT)
pyjwt==2.8.0

# Date/time utilities
python-dateutil==2.8.2

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Type hints
typing-extensions==4.9.0
